+++ constructorio-client/src/main/java/io/constructor/client/models/AutocompleteResponse.java
@@ -1,12 +1,11 @@
 package io.constructor.client.models;
 
 import com.google.gson.annotations.SerializedName;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
 /** Constructor.io Autocomplete Response ... uses Gson/Reflection to load data in */
-public class AutocompleteResponse {
+public class AutocompleteResponse extends ResponseHeaders {
 
     @SerializedName("sections")
     private Map<String, List<Result>> sections;
@@ -16,9 +15,6 @@ public class AutocompleteResponse {
 
     @SerializedName("request")
     private Map<String, Object> request;
-
-    private transient Map<String, List<String>> headers =
-            Collections.<String, List<String>>emptyMap();
 
     /**
      * @return the resultId
@@ -41,13 +37,6 @@ public class AutocompleteResponse {
         return request;
     }
 
-    /**
-     * @return the HTTP response headers
-     */
-    public Map<String, List<String>> getHeaders() {
-        return headers;
-    }
-
     public void setSections(Map<String, List<Result>> sections) {
         this.sections = sections;
     }
@@ -58,9 +47,5 @@ public class AutocompleteResponse {
 
     public void setRequest(Map<String, Object> request) {
         this.request = request;
-    }
-
-    public void setHeaders(Map<String, List<String>> headers) {
-        this.headers = (headers != null) ? headers : Collections.<String, List<String>>emptyMap();
     }
 }
