+++ constructorio-client/src/main/java/io/constructor/client/models/SearchResponse.java
@@ -1,12 +1,10 @@
 package io.constructor.client.models;
 
 import com.google.gson.annotations.SerializedName;
-import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 
 /** Constructor.io Search Response ... uses Gson/Reflection to load data in */
-public class SearchResponse {
+public class SearchResponse extends ResponseHeaders {
 
     @SerializedName("result_id")
     private String resultId;
@@ -16,9 +14,6 @@ public class SearchResponse {
 
     @SerializedName("request")
     private Map<String, Object> request;
-
-    private transient Map<String, List<String>> headers =
-            Collections.<String, List<String>>emptyMap();
 
     /**
      * @return the resultId
@@ -41,13 +36,6 @@ public class SearchResponse {
         return request;
     }
 
-    /**
-     * @return the HTTP response headers
-     */
-    public Map<String, List<String>> getHeaders() {
-        return headers;
-    }
-
     public void setResultId(String resultId) {
         this.resultId = resultId;
     }
@@ -58,9 +46,5 @@ public class SearchResponse {
 
     public void setRequest(Map<String, Object> request) {
         this.request = request;
-    }
-
-    public void setHeaders(Map<String, List<String>> headers) {
-        this.headers = (headers != null) ? headers : Collections.<String, List<String>>emptyMap();
     }
 }
